# Do not recompute twice the whnf of terms in conversion.