Once again, the commit speaks for itself:

> This performance bug was introduced 9 years ago in a8b0345, where the
responsibility of normalizing the term went from ccnv to eqappr in Reduction.
As a result, all recursive calls to eqappr that were preemptively reducing the
term ended up calling whd_stack twice, once by themselves, and once in the
subsequent call to eqappr.
>
> This caused an important slowdown for conversion-intensive proofs, as the
whd_stack calls CClosure.zip to perfom in-place term sharing, leading to
useless huge re-allocations and repetitive write barriers.
>
> Now that eqappr always head-normalizes the term beforehand, we simply don't
call whd_stack anymore when jumping to eqappr.

This patch is a real bug fix, so I'm pretty sure it deserves being merged in 8.6, even though it was introduced 9 years ago! I did not test it thoroughly, but I got up to 30% gain on specific files from UniMath like `UniMath/Combinatorics/Tests.v` w.r.t. the kernel optimization pull-request branch (so that it's 10% then 30%). I'm thus expecting big gains on conversion-intensive files, but I'm waiting for the corresponding infrastructure to assess the exact gain.