@silene 2017-02-27
> I'm thus expecting big gains on conversion-intensive files

While I tend to agree with you, it is not as obvious as you make it sound. Since `whd` has side effects, I am pretty sure one can build testcases where performance are much worse after applying your patch. One just has to have some sharing between terms and make sure that the oracle tells the rhs should be unfolded. Before you patch, the sequence of normalization would have been 2 - 1 - 2; after your patch, it is 1 - 2 - 1 - 2, due to `whd_both` having to perform a recursive call.

So your patch is a guaranteed improvement only for the `whd` calls you have removed from the lhs terms. If you wanted your patch to also be an improvement for the rhs ones, you would have to pass a Boolean to tell `whd_both` which term it should normalize first.