@ppedrot 2017-02-27
> While I tend to agree with you, it is not as obvious as you make it sound. Since whd has side effects, I am pretty sure one can build testcases where performance are much worse after applying your patch. One just has to have some sharing between terms and make sure that the oracle tells the rhs should be unfolded. Before you patch, the sequence of normalization would have been 2 - 1 - 2; after your patch, it is 1 - 2 - 1 - 2, due to whd_both having to perform a recursive call.

In theory I'd definitely agree, but that's not even clear in practice. The more I look at the sharing algorithm in the kernel, the more I'm convinced it is utterly broken. Amongst other weirdnesses, it keeps updating terms even though the substituend is not in normal form (which happens like 99.9% of the time in UniMath) so that you can get strange, hard to predict interactions in the reduction, potentially defeating sharing. Therefore, I think it would be more reasonable to check the effect of this patch on actual code. Note that I'd be still interested in a concrete term triggering the behaviour you're describing!