@ppedrot 2017-02-27
Wait wait wait. I've probably been looking to much at Rust lately, but it seems to me that as of today, there is no possible aliasing between the two terms fed to `eqappr`. The only entry point to that function is `clos_gen_conv` which calls `inject` to turn a `constr` into a `fconstr`, so that no references are shared between the two translated terms. Furthermore, I don't see where in `eqappr` sharing could be introduced between both terms. Did I miss something somewhere? (Note to self: we should really reimplement the kernel in Rust!)

If this is true, then the `whd_both` function is of no use anymore. Maybe at the time aliasing was possible, but this *feature* was removed in between?