@silene 2017-02-27
> I'd be still interested in a concrete term triggering the behaviour you're describing!

I am a bit confused too. I cannot imagine a case where a first term would be in whnf, but putting a second term into whnf would put the first term out of whnf. It would mean that the (already normal) head of the first term was substituted by some other term, but I fail to see how the whnf reduction of the second term could cause that (for instance, no unfolding is performed).