@ppedrot 2017-02-27
We've been discussing by mail about this issue with @maximedenes. They're working on it with @gares and @psteckler, and we're probably going to discuss this topic together at the next workgroup (March 6-7).