@barras 2017-03-15
PMP please:
- when claiming 30% improvements, could you disclose what exactly you measure so we can figure out what it means and how it will affect the user experience
- I wouldn't call a feature "utterly broken" just because it is not optimal. It better applies to "optimizations" that make the code run faster but fail to follow the spec. 
- There *may* be sharing between the lhs and rhs of a conv problem: you just overlooked the clos_infos structure.
- whd_both was introduced to fix an actual bug of the algorithm (as you noticed, it is related to the sharing between both sides of the problem), so removing it will for sure be a problem.

So: this PR changes the strategy of conversion so you can expect dramatic slowdowns in some specific cases, that may be not tested by our benches.