@ppedrot 2017-03-15
> when claiming 30% improvements, could you disclose what exactly you measure so we can figure out what it means and how it will affect the user experience

That's a 30% speedup on user time as displayed by `/usr/bin/time`. I know it is not reproducible and all, but my machine was essentially only processing the Coq file, and it is clearly above the noise threshold.

> I wouldn't call a feature "utterly broken" just because it is not optimal. It better applies to "optimizations" that make the code run faster but fail to follow the spec.

Right, that may be an overstatement of mine uttered in the heat of the perfing rush. At least, it is not performing call-by-need as I know it (i.e. memoizing computations that are potentially duplicated), it is doing fishy things like repeatedly updating the same cell with non-normalized terms, and experimentally, it is doing reaaaaaally bad (like half of the time passed zipping terms for nothing). I did not find in your PhD a theoretical justification for the specific memoization pattern implemented in Coq kernel by the way, apart from an upper bound on the number of in-place update. Do you have more material somewhere?

> There may be sharing between the lhs and rhs of a conv problem: you just overlooked the clos_infos structure. whd_both was introduced to fix an actual bug of the algorithm (as you noticed, it is related to the sharing between both sides of the problem), so removing it will for sure be a problem.

Yes, I overlooked this and realized it afterwards. I was wondering whether a simple, stupid solution would be to simply have a distinct clos_info cache for each side of the conversion arguments.

> So: this PR changes the strategy of conversion so you can expect dramatic slowdowns in some specific cases, that may be not tested by our benches.

True, but as for all heuristics we should see whether this affects practical instances of reduction. Since we have the VM and native_compute, it is not clear at all whether the call-by-value-like computations that one wanted to perform at the time are still relevant.