@ejgallego 2017-03-25
So what's the status on this, I guess not a candidate for 8.6.1, right? Mabye @matejkosik should add this to the queue?