@matejkosik 2017-03-27
@ejgallego 
> Maybe @matejkosik should add this to the bench queue?

[Measurements](https://ci.inria.fr/coq/view/benchmarking/job/benchmark-the-whole-branch/81/console)  (from 3 iterations we get precision around 1%):
```
┌────────────────────────┬─────────────────────────┬─────────────────────────────────────┬───────────────────────────────────────┐
│                        │      user time [s]      │             CPU cycles              │           CPU instructions            │
│                        │                         │                                     │                                       │
│           package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │            NEW            OLD PDIFF   │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│            coq-unimath │ 1206.74 1289.22 -6.40 % │ 3346073827567 3574313500412 -6.39 % │  5615506523242  5966960513460 -5.89 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│   coq-mathcomp-algebra │  170.09  172.21 -1.23 % │  471343759996  478003529638 -1.39 % │   673396529582   682661902991 -1.36 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│  coq-mathcomp-fingroup │   57.78   58.30 -0.89 % │  159372150518  160628041381 -0.78 % │   224367700110   225858418038 -0.66 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-ssreflect │   43.32   43.67 -0.80 % │  118574724646  118866642627 -0.25 % │   151832588954   152227611684 -0.26 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│               coq-corn │ 1524.37 1535.87 -0.75 % │ 4233098772873 4268773901158 -0.84 % │  6798613456209  6849887836684 -0.75 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-character │  265.34  267.24 -0.71 % │  738160083039  743240663605 -0.68 % │  1097901003545  1107320575693 -0.85 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│        coq-fiat-crypto │ 2305.82 2319.46 -0.59 % │ 6414660916396 6450680908397 -0.56 % │ 10846102878418 10877650312583 -0.29 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│  coq-mathcomp-solvable │  199.54  200.45 -0.45 % │  554165358528  556525773267 -0.42 % │   827179405886   831229952320 -0.49 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│       coq-fiat-parsers │  513.01  515.07 -0.40 % │ 1407528311508 1411490285744 -0.28 % │  2015098914828  2024413307479 -0.46 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│               coq-iris │  426.53  428.17 -0.38 % │ 1187000598620 1187549205459 -0.05 % │  1721897368355  1723377121214 -0.09 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│     coq-mathcomp-field │  480.86  482.07 -0.25 % │ 1338605850811 1341639687247 -0.23 % │  2256930876086  2266162412033 -0.41 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│                 coq-sf │   44.47   44.58 -0.25 % │  121934115338  121978611309 -0.04 % │   159940539914   160122672033 -0.11 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│               coq-hott │  228.86  229.38 -0.23 % │  618038173840  619916088477 -0.30 % │  1040847042063  1043420252931 -0.25 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│              coq-color │  635.49  635.86 -0.06 % │ 1762594095901 1761749850681 +0.05 % │  2287808491839  2290683962279 -0.13 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│       coq-math-classes │  221.40  221.37 +0.01 % │  610644499890  610484428036 +0.03 % │   862007249587   862686497112 -0.08 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│         coq-coquelicot │   70.07   70.06 +0.01 % │  193897962568  194466103015 -0.29 % │   245212232007   245138446031 +0.03 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│           coq-compcert │  818.02  817.83 +0.02 % │ 2273201544496 2273908111386 -0.03 % │  3493417396336  3498122320024 -0.13 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│             coq-geocoq │ 2402.59 2401.32 +0.05 % │ 6690155823115 6688265170874 +0.03 % │ 11360086769029 11375921231190 -0.14 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│              coq-flocq │   51.86   51.72 +0.27 % │  143295159194  143196838958 +0.07 % │   183917399724   183992293845 -0.04 % │
└────────────────────────┴─────────────────────────┴─────────────────────────────────────┴───────────────────────────────────────┘


PDIFF = proportional difference between measurements done for the NEW and the OLD Coq version
      = (NEW_measurement - OLD_measurement) / OLD_measurement * 100%

NEW = 256ca51bafc7200c8c006981cad60e57014e0dbc
OLD = 27e8d8857ea5435ccec9eddd6c34324de82afd32
```